+++ src/components/Archive/list.js
@@ -1,6 +1,7 @@
 import React from "react"
 import styled from "styled-components"
-import { useStaticQuery, graphql, Link } from "gatsby"
+import { useStaticQuery, graphql } from "gatsby"
+import ListItem from "./listItem"
 
 const Wrapper = styled.aside`
   height: 100%;
@@ -10,8 +11,8 @@ const Wrapper = styled.aside`
   display: flex;
   flex-direction: column;
   background-color: #141621;
-  padding: 10px;
   text-align: center;
+  padding-bottom: 12px;
 
   @media screen and (max-width: 1200px) {
     width: 100%;
@@ -20,28 +21,20 @@ const Wrapper = styled.aside`
 `
 
 const Heading = styled.h1`
-  margin: 0.5em 0;
+  height: 100px;
+  line-height: 100px;
   color: #ff5370;
-  font-size: 2.5em;
+  font-size: 2em;
   font-weight: 800;
-  display: inline-block;
   width: 100%;
+  margin: 0;
 
   &::after {
     border-top: 2px solid #fff;
     display: block;
-    width: 33.3%;
+    width: 20%;
     content: "";
-    margin: 4px auto 0;
-  }
-`
-
-const ListItem = styled.div`
-  padding: 12px 10px;
-  border-bottom: 1px solid #ff5370;
-
-  &:last-of-type {
-    border-bottom: none;
+    margin: -26px auto 0;
   }
 `
 
@@ -51,20 +44,7 @@ const ListContainer = styled.div`
   flex: 1;
 `
 
-const ItemHeading = styled(Link)`
-  display: block;
-  cursor: pointer;
-  font-size: 1.1em;
-  text-decoration: none;
-  color: white;
-  margin: 0;
-
-  :hover {
-    text-decoration: underline;
-  }
-`
-
-const List = () => {
+const List = ({ episode: currentEpisode }) => {
   const data = useStaticQuery(graphql`
     query EpisodeList {
       allFile(
@@ -97,16 +77,14 @@ const List = () => {
       <ListContainer>
         {episodes
           .map((episode, i) => (
-            <ListItem key={episode.id}>
-              <p style={{ color: "#ffffff88", margin: 0, fontSize: "0.9em" }}>
-                {episode.childMarkdownRemark.frontmatter.publicationDate}
-              </p>
-              <ItemHeading
-                to={`/archive${episode.childMarkdownRemark.frontmatter.slug}`}
-              >
-                {episode.childMarkdownRemark.frontmatter.title}
-              </ItemHeading>
-            </ListItem>
+            <ListItem
+              key={i}
+              episode={episode}
+              current={
+                episode.childMarkdownRemark.frontmatter.slug ===
+                currentEpisode.frontmatter.slug
+              }
+            />
           ))
           .reverse()}
       </ListContainer>
