+++ src/components/Player/volumeBars.js
